# api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: Workflow orchestration platform for financial validation
  termsOfService: http://swagger.io/terms/
  title: Flowker API
  version: 1.0.0
servers:
- url: //localhost:4021/
paths:
...
  /v1/dashboards/executions:
    get:
      description: Retrieves an aggregated summary of executions with optional time
        range and status filters
      parameters:
      - description: Start time filter (RFC3339)
        in: query
        name: startTime
        schema:
          format: date-time
          type: string
      - description: End time filter (RFC3339)
        in: query
        name: endTime
        schema:
          format: date-time
          type: string
      - description: Status filter
        in: query
        name: status
        schema:
          enum:
          - pending
          - running
          - completed
          - failed
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ExecutionSummaryOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Get execution summary
      tags:
      - Dashboard
...
  /v1/executions:
    get:
      description: Retrieves a paginated list of workflow executions with optional
        filters
      parameters:
      - description: Filter by workflow ID
        in: query
        name: workflowId
        schema:
          format: uuid
          type: string
      - description: Filter by status
        in: query
        name: status
        schema:
          enum:
          - pending
          - running
          - completed
          - failed
          type: string
      - description: Page size (1-100)
        in: query
        name: limit
        schema:
          default: 10
          maximum: 100
          minimum: 1
          type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        schema:
          type: string
      - description: Sort field
        in: query
        name: sortBy
        schema:
          default: startedAt
          enum:
          - startedAt
          - completedAt
          type: string
      - description: Sort direction
        in: query
        name: sortOrder
        schema:
          default: DESC
          enum:
          - ASC
          - DESC
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ExecutionListOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: List executions
      tags:
      - Executions
  /v1/executions/{id}:
    get:
      description: Retrieves the current status of a workflow execution
      parameters:
      - description: Execution ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ExecutionStatusOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Get execution status
      tags:
      - Executions
  /v1/executions/{id}/results:
    get:
      description: Retrieves the results of a completed workflow execution
      parameters:
      - description: Execution ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ExecutionResultsOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Execution still in progress
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Get execution results
      tags:
      - Executions
...
  /v1/workflows/{workflowId}/executions:
    post:
      description: Starts a new workflow execution. Returns immediately with pending
        status.
      parameters:
      - description: Workflow ID
        in: path
        name: workflowId
        required: true
        schema:
          format: uuid
          type: string
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ExecuteWorkflowInput'
        description: Execution input
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ExecutionCreateOutput'
          description: Idempotent request
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ExecutionCreateOutput'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Conflict
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Payload too large (max 1 MB)
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Execute a workflow
      tags:
      - Executions
      x-codegen-request-body-name: execution
components:
  schemas:
    api.ErrorResponse:
      properties:
        code:
          example: FLK-0001
          type: string
        message:
          example: Invalid input provided
          type: string
        title:
          example: Bad Request
          type: string
      required:
      - code
      - message
      - title
      type: object
...
    github_com_LerianStudio_flowker_pkg_model.ExecuteWorkflowInput:
      properties:
        inputData:
          additionalProperties:
            type: object
          type: object
      required:
      - inputData
      type: object
    github_com_LerianStudio_flowker_pkg_model.ExecutionCreateOutput:
      example:
        executionId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        startedAt: startedAt
        workflowId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: status
      properties:
        executionId:
          format: uuid
          type: string
        startedAt:
          type: string
        status:
          type: string
        workflowId:
          format: uuid
          type: string
      type: object
    github_com_LerianStudio_flowker_pkg_model.ExecutionListOutput:
      example:
        nextCursor: nextCursor
        hasMore: true
        items:
        - executionId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          completedAt: completedAt
          currentStepNumber: 0
          errorMessage: errorMessage
          totalSteps: 6
          startedAt: startedAt
          workflowId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: status
        - executionId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          completedAt: completedAt
          currentStepNumber: 0
          errorMessage: errorMessage
          totalSteps: 6
          startedAt: startedAt
          workflowId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: status
      properties:
        hasMore:
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ExecutionStatusOutput'
          type: array
        nextCursor:
          type: string
      type: object
    github_com_LerianStudio_flowker_pkg_model.ExecutionResultsOutput:
      example:
        stepResults:
        - output:
            key: '{}'
          executedAt: executedAt
          stepName: stepName
          errorMessage: errorMessage
          stepNumber: 6
          durationMs: 0
          nodeId: nodeId
          status: status
        - output:
            key: '{}'
          executedAt: executedAt
          stepName: stepName
          errorMessage: errorMessage
          stepNumber: 6
          durationMs: 0
          nodeId: nodeId
          status: status
        executionId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        completedAt: completedAt
        finalOutput:
          key: '{}'
        startedAt: startedAt
        workflowId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: status
      properties:
        completedAt:
          type: string
        executionId:
          format: uuid
          type: string
        finalOutput:
          additionalProperties:
            type: object
          type: object
        startedAt:
          type: string
        status:
          type: string
        stepResults:
          items:
            $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.StepResultOutput'
          type: array
        workflowId:
          format: uuid
          type: string
      type: object
    github_com_LerianStudio_flowker_pkg_model.ExecutionStatusOutput:
      example:
        executionId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        completedAt: completedAt
        currentStepNumber: 0
        errorMessage: errorMessage
        totalSteps: 6
        startedAt: startedAt
        workflowId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: status
      properties:
        completedAt:
          type: string
        currentStepNumber:
          type: integer
        errorMessage:
          type: string
        executionId:
          format: uuid
          type: string
        startedAt:
          type: string
        status:
          type: string
        totalSteps:
          type: integer
        workflowId:
          format: uuid
          type: string
      type: object
    github_com_LerianStudio_flowker_pkg_model.ExecutionSummaryOutput:
      example:
        running: 2
        total: 150
        pending: 3
        completed: 120
        failed: 25
      properties:
        completed:
          example: 120
          type: integer
        failed:
          example: 25
          type: integer
        pending:
          example: 3
          type: integer
        running:
          example: 2
          type: integer
        total:
          example: 150
          type: integer
      type: object
...
    github_com_LerianStudio_flowker_pkg_model.StepResultOutput:
      example:
        output:
          key: '{}'
        executedAt: executedAt
        stepName: stepName
        errorMessage: errorMessage
        stepNumber: 6
        durationMs: 0
        nodeId: nodeId
        status: status
      properties:
        durationMs:
          type: integer
        errorMessage:
          type: string
        executedAt:
          type: string
        nodeId:
          type: string
        output:
          additionalProperties:
            type: object
          type: object
        status:
          type: string
        stepName:
          type: string
        stepNumber:
          type: integer
      type: object
...